update _check_error & rename qianfan_requestor
