+++ src/qianfan/resources/tools/utils.py
@@ -20,7 +20,7 @@ from typing import Any, Awaitable, Callable
 
 from qianfan.consts import DefaultValue
 from qianfan.resources.auth.oauth import Auth
-from qianfan.resources.requestor.qianfan_requestor import QfAPIRequestor
+from qianfan.resources.requestor.openapi_requestor import QfAPIRequestor
 from qianfan.resources.typing import ParamSpec, QfRequest, QfResponse, RetryConfig
 
 P = ParamSpec("P")
